WriterContainer -> FileFormatWriter
